+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 7.0.0
+
+* Use Amazon Linux 2003 and Python 3
+* Added example/ubuntu as a test to create an installation of Docker for development using Ubuntu and have capability of using the latest version of Docker.
+
 ## 6.0.3
 
 * Removed the one instance of `template_file` that is triggering the load of the Hashicorp template module.